added installation section into the lisp doc
This was gerrit 11303 before it was abandoned. I did my best to recreate
the changes that were in 11303 before the patch it was dependent on was
abandoned. Feel free to merge this or tell me to abandon it as well.

Based on a patch by Divya Shetty <dshetty@brocade.com>

Change-Id: Iee47eb9c009ec2faf40f74d322645d31629a20f9
Signed-off-by: Divya Shetty <dshetty@brocade.com>
Signed-off-by: Colin Dixon <colin@colindixon.com>