Update patch set 2
Patch Set 2: Code-Review+1

I'm not really familiar with the documentation generation and format, so forgive me if this is a dumb question, but is there a way to look at this in some other format?  I tried checking out and building, but all I found was an xml file and the associated jpg files.

Patch-set: 2
Reviewer: Gerrit User 226 <226@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=+1