Fix site deploy ID
The deploy ID for sites shouldn't be dynamically setting the name to
${project.artifactId} as this makes it difficult to properly configure
the credentials on the Jenkins systems. This change sets it to something
similar to the release and snapshot repositories and therefore easier to
script in the puppet manifests that build the credential files.

This should resolve the issue of not being able to deploy generated
sites to Nexus.

Change-Id: I5706a3da95702fc6f36ed531d134583a5e9a489d
Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>