Update patch set 1
Patch Set 1: Code-Review-1

Hi,

The document outlines for both the dev guide and the user guide are empty.

Please take a moment to type in the headings for each document, so that you have a rough table of contents.

Also, an overview of the topic would be appreciated.

Thanks,
Moi

Patch-set: 1
Reviewer: Gerrit User 2595 <2595@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=-1