Update patch set 6
Patch Set 6:

This looks good overall. My only question is what sections does SFC want to have as the top-level sections in the user guide and developer guide.

I'm honestly curious. The goal is to make it the most useful for users and developers. As I said in previous feedback, one way to do it would be to cluster it all under one global "SFC" feature group, but if there are logical separations for users or for developers that make sense, we should give them.

Patch-set: 6
Label: Verified=0