Update patch set 1
Patch Set 1: Code-Review-1

(3 comments)

This mostly looks good. The only thing I'd ask is to replace the verbaitm text from the templates with either TBD or, bettter, a sentence or two. If that's going to be immediately done by other groups with other patches, I could probably be convinced to merge this anyway.

Patch-set: 1
Label: Code-Review=-1