Update patch set 3
Patch Set 3: Code-Review-1

(15 comments)

Along with the inline comments, I think we also want to make sure to say explicitly that we are assuming that people are deploying OpenDaylight on a secured management network and thus the controller will not be subject to wanton attack from either the data network or the Internet at large.

I'm also going to ask for feedback on this patch from the security response team.

Patch-set: 3
Label: Code-Review=-1