Update patch set 1
Patch Set 1: Code-Review-1

It's very hard to tell if the comments from patch 20762 were addressed when this is a separate patch. Is it possible to submit this as patch set 3 of change 20762 instead?

The simplest way to do that would be to
1.) Restore 20762 from being abandoned
2.) checkout this change
3.) do a git commit --amend and replace the changeId with the one from 20762 
4.) push/git review again
5.) abandon this change

Patch-set: 1
Reviewer: Gerrit User 48 <48@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=-1