Update patch set 3
Patch Set 3: Code-Review-1

First, I updated the patch so that (i) the files are in the right directory since we renamed the installation guide to the getting started guide and (ii) the files are actually included in the manual.

Second, did we want this content in the getting started guide or the user guide? If we want it in the getting started guide, please also remove it from the user guide as part of this patch.

Patch-set: 3
Label: Code-Review=-1