Update patch set 2
Patch Set 2:

In patch set 2, I reworded the introductory text and the Message Order Preservation section a bit. I also removed the section about important dependent distributions because it was aimed at contributors to the OpenFlow plugin instead of developers using the OpenFlow plugin, while this guide is supposed to be aimed at the latter.

If these changes look good, please +1 this, I'll merge it and keep working on the later contents.

Patch-set: 2
Label: Verified=0