Update patch set 4
Patch Set 4: Code-Review-1

(5 comments)

First, the new Cassandra content wasn't linked into the main docs. 

Second, it seems like (at least for the install guide) that the content is nearly identical between the three different chapters and that maybe they could be mostly merged with just the different parts called out in the right place? Does that seem reasonable?

Third, there are a few small comments about the factual accuracy around some of the text. In general, it might be a good idea to avoid talking about the particular release, e.g., Beryllium or Lithium, unless it adds value to the reader as it has to be updated carefully.

I'm going to upload a new patch set in a few seconds that fixes some of this. You can pick up from there using these instructions:
https://wiki.opendaylight.org/view/Documentation/Tools#Updating_an_Existing_Patch

Patch-set: 4
Label: Code-Review=-1