Update patch set 1
Patch Set 1: Code-Review-1

(4 comments)

It seems like the getting started guide content is really also developer content and that there isn't any user-facing part of NeXt itself. Is that right? If so, could we fix that here.

I haven't done an in-depth review of the content, but it looks like it will need to be proofread at least once. It would be faster for me to review if you could have somebody do that rather than me having to do it later, but I'll do it if not.

Patch-set: 1
Reviewer: Gerrit User 48 <48@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=-1