Update patch set 8
Patch Set 8: Code-Review-1

(18 comments)

I haven't had a chance to review the text in the how tos yet, but there are some comments in the other sections.

The high-order bit is that I try to avoid saying "OpenDaylight Controller" because it can be confusing whether you mean the controller project inside OpenDaylight or the entirety of OpenDaylight. Instead I just say OpenDaylight, not "the OpenDaylight controller" or "ODC" or "ODL".

Patch-set: 8
Label: Code-Review=-1