Update patch set 6
Patch Set 6:

I made some small changes to the information here. Mostly removing the cloning/building with maven since that's not what a user should be doing and then changing line-ends from windows to UNIX.

If you like the changes, I'll merge this, cherry-pick it to stable/beryllium and merge it there.

Patch-set: 6