Updated git submodules
Project: releng/builder master 4995838d169f87cb135a3425f6bcc8f765e59b9b

Rearrange the user creation step for slaves

With the change from JClouds to the OpenStack slave plugin we need to
fully manage the creation of the Jenkins user. This modifies the timing
of when the user gets created to be the last step. This will hopefully
prevent us from having issues with the init scripts not having been
completed before a job starts on the system.

Secondly, we handle adding Jenkins to the docker group in a slightly
more sane manner but without breaking our current production setup that
executes some, but not all of these scripts and in a different order.

Change-Id: I8d88cd0d9de6b3defd6615faa474e4ce8d583ea4
Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>