Updated git submodules
Project: releng/builder master 3d7573a98a4cf47d801735c6b37209827204dbca

Merge "Fix archiving script to trim the end of the log file"

Fix archiving script to trim the end of the log file

Because the console log is still running while we are downloading it to
archive to logs.opendaylight.org it ends abruptly and causes confusion.
This patch trims the log so it will have a natural end to it's log.

Change-Id: Ic68828d2209a1c95e3c8282a82bc12f3138105a6
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>