Update patch set 7
Patch Set 7: Code-Review+1

Patch sets 4-7 are just minor grammar and formatting changes.

This is clearly Beryllium focused (and I'm cherry-picking it to stable/beryllium now), but it looks good and I'm happy to merge it. It would be good to make another pass and change the Beryllium stuff to Boron: mostly a few URLs and references to Beryllium.

If you (Tony or Tomas) like it, please +1 and I'll merge it.

Patch-set: 7