Disable javadoc generation
This is a follow-on to 86f036f5b009706a5cf1e0ea61c3aae2f0467f21, which
seems to have not quite fully-disabled javadoc generation.

ReadTheDocs has a build timeout limit of 902 seconds. JavaDoc generation
adds a significant amount of time to our build so disable it until we
have a solution for this otherwise our docs won't get updated on
readthedocs.

Change-Id: Ifc98387a066a4a60b3676951adf142fe2071b803
Signed-off-by: Colin Dixon <colin@colindixon.com>