Update patch set 1
Patch Set 1: Code-Review-1

(27 comments)

Lots of grammar and formatting issues. It's not really clear what you're trying to convey in places, so I asked for clarification. The formatting is a bit off in places. I'd suggest you build the documentation and look at the built PDFs to get an idea of how it's formatting. See here for how to do that:
https://wiki.opendaylight.org/view/Documentation/Tools#Generating_Documentation

Patch-set: 1
Label: Code-Review=-1