Update patch set 1
Patch Set 1:

(I tried to quote all of the comment, but it didn't work for some reason.)

In my next commit, I've changed every occurrence to "YANG IDE", and all contractions are expanded.

I've always been OCD frustrated with Maven's "group id" and "artifact id", as the actual dialog fields and labels seem to randomly vary.  In this doc, I've been very careful to ensure that if the string is in quotes, it represents exactly the spelling you would see on the screen.  If I refer to the concept in a sentence, I try to use "group id", because, like, that's what it is.  Attempting to make these "consistent" in this case won't increase the understanding of the content, and will result in using words in an English statement that aren't actually words.

Added blank lines above each of the list blocks.

Completed mentioned quoted string.

I'm hesitant to mention the OpenDaylight formatting of the group id, as I'm trying to reduce any possible cognitive load on users who don't necessarily use OpenDaylight (a contentious issue in AT&T).

Removed the mention of the content in an upcoming section.

And fixed a typo neither of us saw before. :)

Patch-set: 1
Reviewer: Gerrit User 3576 <3576@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0