Update patch set 1
Patch Set 1:

Is this supposed to be a new version of this: https://git.opendaylight.org/gerrit/#/c/39165/

I think so. If so, could you abandon this patch and upload it as a new patch set there. You can do that by doing:

git commit --amend

on top of this patch and setting the Change-Id to match the one in 39165.

Patch-set: 1
Reviewer: Gerrit User 48 <48@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0