Updated git submodules
Project: integration/test master 36c709e784d0c33716c6d742e66817f0b08feb59

Adapt 404 response when delete non-existent element

Add new function "Remove All Elements at URI If Exists" inside Utils.robot
which sends a Delete REST request only if the element sent as input
parameter exists in the datastore.

Delete all calls to "Remove All Elements at URI" at the beggining of every
test case, which was used as "test case setup" and add a Robot FW Test Case
Setup calling the function "Remove All Elements at URI If Exists".

Adapt test cases verifying the deletion of a non-existent element to avoid
calling "Remove All Elements at URI". This function checks that the
result code returned is (200, 201 or 204) and since boron release a 404 is
returned if the element to be removed does not exist in the datastore.
Instead, a delete request is sent and it is checked that the result
obtained is 404 when the element is not present.

Change-Id: Icfba27e646a19aa27f1be24afc398c87b7323ebf
Signed-off-by: Jose Luis Franco Arza <jose.luis.franco.arza@ericsson.com>