Update patch set 4
Patch Set 4: Code-Review-1

I'm going to push a patch that fixes this, but I'm exceedingly disappointed at how little effort went in to this documentation. The developer guide is *still* not included properly and the zero effort seems to have been put into migrating the formatting from the wiki.

Patch-set: 4
Label: Code-Review=-1