Update patch set 5
Patch Set 5: Code-Review-1

(3 comments)

Adding Thanh as a reviewer here as well. My guess is that the logs.opendaylight.org links are going to expire and that the right approach here is to publish this information as a maven site, but I could be wrong.

Thanh, could you help? Basically they were linking to lastSuccessfulBuild/artifacts in jenkins which no longer work.

Hideyuki-san and YOGA, it might make sense to fix everything but those {jenkins,logs}.opendaylight.org links so we can merge this patch and then fix those in a follow-up since I think the fix is likely to involve more than just fixing the docs.

Patch-set: 5
Label: Code-Review=-1