Updated git submodules
Project: odlparent master d5aecae24fb4204513a96c4e632ca85dbc41f81a

Have slf4j-simple as an "always there" <scope>test dependency.

The "always there" as in always, literally, not just in
<dependencyManagement>.  This makes sense because in (unit and
component, not integration) tests you want logging to "just work" - you
don't want to have to remember to add slf4j-simple yourself as a
dependency every time (and you certainly don't want to have to add
logback every time, and configure it with an XML or properties; for
every project).

Also the problem with manually adding a slf4j-simple dependency to
projects with tests is that there is a risk of it getting accidentally
removed, because "mvn dependency:analyze" will find it's "not needed";
as happened e.g. in
https://git.opendaylight.org/gerrit/#/c/45090/6/vpnservice/aclservice/impl/pom.xml

Change-Id: I5af53bbd8d69ad5cebac08f97225c2f9cbe60ab7
Signed-off-by: Michael Vorburger <vorburger@redhat.com>