Updated git submodules
Project: releng/builder master 4da04073ef0bb40c34833941480bf3fcb8979506

Merge "Preserve property files to trigger test jobs"

Preserve property files to trigger test jobs

The tigger to the ingetration-test jobs are in the publisher
section, while the workspace gets cleaned up at the end of builders
resulting in downstream jobs to fail, as the property file is not found
once the workspace is cleaned up.

As a workaround we have to preserve the variables.prop file for the
autorelease jobs to tigger the downstream jobs, which is accomplished
excluding the file.

Change-Id: I7c023bcb70ea8f4ec4bd9cf648e37a5748ee111c
Signed-off-by: Anil Belur <abelur@linuxfoundation.org>