Updated git submodules
Project: releng/builder master 548b2b595de4b433d3e1b6df47365e1c64b15454

Redirect zip logs to file and store on logs.o.o

The zip logs take a lot of the console log lines at the end of a build.
The output might be useful for troubleshooting so lets redirect the
output to a file and store it in logs.opendaylight.org instead of
outputting it to the Jenkins console logs directly.

Change-Id: Ie56d2a5348c9a55854199539e647f21894febe68
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>