Updated git submodules
Project: releng/builder master 6b397bce33c029681e8132547a603080fbf430f3

Merge "Force Jenkins to not change it's votes for merge"

Force Jenkins to not change it's votes for merge

Jenkins by default resets the votes it has in Gerrit to 0. Gerrit 2.13 has
problems with this and causes us to catch an error. As such, we should make
sure that Jenkins does not try to change its votes when it starts during a
merge job. We assume that any merge run will be with our normal mechanically
set votes.

Change-Id: Id7aad85b0d87378660e017f20a0efe35b0a39dff
Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>