Fix submodule tracking branch
For aaa, netconf, and odlparent we should be tracking stable/boron in
the stable/boron docs branch and not master. This patch fixes these
submodules to start tracking the right boron branch.

According to upstream Gerrit setting a branch specifier of "." means to
track the same branch as the super project's branch.

Change-Id: Ifc5a0d1487008d61f1ca0184687eec04ce00b8e9
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>