Set branch specifier to "."
According to upstream Gerrit setting a branch specifier of "." means
that the submodules will track the same branch as the super project's
branch as long as those branches exist in the submodule.

This future proofs our .gitmodules so when we perform branch cutting we
don't have to remember to update .gitmodules.

Change-Id: I801550ef32786f26a83f2282eada0b990fa91258
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>