Update patch set 1
Patch Set 1:

Just realized once I started writing this that it wasn't about initializing in project repo docs but how to contribute project documentation to the user and developer guides.

I will make the corrections but wanted to ask if it's worth also adding details on how to put the docs in the project repos and link them?

Or perhaps maybe that would be better as a follow up patch and have this patch purely focus on converting the Wiki page to reStructuredText.

Patch-set: 1
Label: Verified=0