Update patch set 1
Patch Set 1:

This looks good to me too. Thanks for merging it Thanh. Just for posterity's sake, this was self-merged because it needs to be merged before any patches are merged to the submodule project in order for it to automatically track it going forward.

Patch-set: 1
Reviewer: Gerrit User 48 <48@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0