Update patch set 4
Patch Set 4:

(5 comments)

The WIP is here until we are sure about the content of the tutorial. But if you can give any review on grammatical or spelling aspects of the patch, that would be so great. Once the WIP is removed, to the best of my knowledge, that should indicate the patch is completely ready.

Patch-set: 4
Reviewer: Gerrit User 6491 <6491@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0