Update patch set 5
Patch Set 5:

> (8 comments)
 > 
 > Generated doc: https://logs.opendaylight.org/releng/jenkins092/docs-verify-rtd-carbon/502/archives/docs/_build/html/opendaylight-with-openstack/openstack-with-netvirt.html
 > 
 > Had a quick pass, mainly from formatting perspective, will revisit
 > content later.
 > 
 > There are some warnings when compiling the doc. Would be good to
 > have zero warnings. You can see them when compiling in your setup
 > or in https://jenkins.opendaylight.org/releng/job/docs-verify-rtd-carbon/502/consoleFull

Thanks Vishal, fixed some of the above but this is definitely a WIP.
Need to get to know rst and styling.

How do I verify "compilation" manually? I saw running tox -edocs but it doesn't seem to work for me locally.
Also is there any "beautify" tool we can run to auto-format?

Regarding toctree, isn't it just for referencing other .rst files? Or can we use it as a table of contents for subheaders?

Patch-set: 5
Label: Verified=0