Update patch set 1
Patch Set 1: Code-Review-1

The only reason this is not a straight cherry-pick of:
https://git.opendaylight.org/gerrit/#/c/49322/

is because the previous patch's cherry-pick hadn't been merged yet. It has now, so I *think* we can just cherry-pick 49322 and abandon this.

If that's not right, let me know. If it is, can you abandon this patch?

Patch-set: 1
Reviewer: Gerrit User 48 <48@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=-1