Updated git submodules
Project: releng/builder master 61648282c19c20fc0d51f2fcde2079bcd79ccbd3

Merge "Switch to build-discarder property for logrotate"

Switch to build-discarder property for logrotate

logrotate is deprecated in Jenkins>=1.637 so switch to the new style
build-discarder property.

This is a first patch to test with the releng project. Follow on patches
will swap this setting for the remaining jobs.

Change-Id: I6b0f15852024e66c8584ee604a555ec320d1b0be
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>