Updated git submodules
Project: releng/builder master f571cb39b33e760f2267691cf97d792d36848882

Merge "Images with pre-loaded devstack pips and yums"

Images with pre-loaded devstack pips and yums

This patch adds templates and a provisioning
script to build images that are pre-installed
with (a lot of) the various pip and yum pkgs
required for running devstack.

This aims to (hopefully) reduce the run time of CSIT.

Note that this is a just a first step. The low hanging
fruit. There are more packages that are installed by
openstack projects that can be preinstalled here as well.

Change-Id: Ic837dd2ef7a96141130bb0d71f578cd06de5f1f4
Signed-off-by: Josh <jhershbe@redhat.com>