Update patch set 3
Patch Set 3:

> > Do we also want to add this to stable/boron? I guessing no.
 > 
 > I also lean towards no. Int/Pack's model is that it sits outside of
 > ODL releases, so Int/Pack master should apply to all current ODL
 > versions.

For better or for worse. That's not how read the docs works. If you want to have it show up for all releases, it needs to be included from all branches. You can include the master branch of int/packaging from all branches so that things are all actually the same. If that makes sense.

Patch-set: 3