Updated git submodules
Project: odlparent master 863bf4b167d899f1e681c46d2c0e2e7cd2311393

Add -Dsft.diag.timeout=... option to SingleFeatureTest BundleDiagInfos

This is useful when locally debugging issues, and you don't want to wait
too long for the bundle convergence to eventually succeed (because you
already know there is a problem).  The worst case default of 5 minutes
for the build is still a good idea, though.

This also tunes the pollInterval from 250ms to 1s, which is sufficient,
and leads to less noise in the log.

This also fixes -Dsft.diag.skip=true, which didn't actually work
(becaues of a missing PropagateSystemPropertyOption).

Change-Id: I8c7bae10295ddef57c0869628789949d991456e3
Signed-off-by: Michael Vorburger <vorburger@redhat.com>