Updated git submodules
Project: aaa master 634060e7c1d036d3b1de26efcdd195d98b1a940e

Merge "Add ODLPrincipal to API so it is exposed for outside use"

Add ODLPrincipal to API so it is exposed for outside use

A recent request in the mailing list was to expose ODLPrincipal
for use from other contexts.  This will also be useful for when
the MDSALRealm is added, which will also utilize ODLPrincipal
constructs.

Change-Id: Ic2b4bc9646f50a6fb2e2e1d43b814d2bc4e8c1f1
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>