Updated git submodules
Project: integration/test master 954517fda58e484e1d5b036e1b16ee8f8241e4c6

Skip working on gerrits if the includes list is empty

this could happen when the firt gerrit found is not "merged" as would
be the case when this tool is triggered with a patch test gate job.

this will actually stop parsing gerrits for the project that has an
unmerged change in the distro.

this is a quick fix to get the tool working again. a better fix needs
to come next where it can handle unmerged gerrits and then continue
to print all the merged patches that still made it in to that relevant
project

Change-Id: I96c683cf07e7ed43c5eab42a48a8a0c8627d04e6
Signed-off-by: Jamo Luhrsen <jluhrsen@redhat.com>