Updated git submodules
Project: genius master 77db03ded577cab3b9359a08421602df8c771519

Made mdsalutil-api IPv4's protocolClassMap private instead of public

I've verified that this public static final (so constant) field does not
actually seem to be used by any code - by simple local grep, so I could
theoretically have missed something, please double check.

The inline TODO comment on etherTypeClassMap makes me think that this
may not be workin in OSGi as originally designed, anyway?   This should
perhaps be re-architected as an OSGi service, which bundles can
contribute new types to?

I've stumbled upon this while working on Checkstyle clean-up (it
complains that this constant should normally be upper-case, like
PROTOCAL_CLASS_MAP, instead of protocolClassMap), so I had a closer
look, and upon finding that it's not actually used outside of this
class, propose to make it private.

If this must stay public for some reason that eludes me, then we would
have to put a @SuppressWarnings("checkstyle:ConstantName") on this (like
I have for Ethernet's etherTypeClassMap in
https://git.opendaylight.org/gerrit/#/c/51390/, just because I saw that
already seems to be used in some places).
 
Change-Id: I16035954840d28e53db852e1f1ca5621f49433b2
Signed-off-by: Michael Vorburger <vorburger@redhat.com>