Updated git submodules
Project: integration/test master 5e02227c1db5c58292cafc178bb053c440280eb5

Process unmerged gerrits in distchanges

gerrits that have not merged yet are not found when doing
git queries because the queries assume they are merged. This
patch first tries to find the gerrit merged and if not
found it tries as not merged.

Change-Id: Ibce154effe9e58b9cad1a6b88df7c0356a90698f
Signed-off-by: Sam Hague <shague@redhat.com>