Updated git submodules
Project: releng/builder master 467f8bc7501fdfb3268c21228148c7081952c6ee

Improve ping output to be less verbose

The full ping output takes a lot of log space and makes the logs hard to
read. We are only interested in knowing what IP it's pinging and if it
was success or fail so configure a custom message.

Change-Id: I5a03963accf9800c1fcd154293e5245d889a7709
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>