Add first version of the AAA Developer Guide
This is the first version of the AAA Developer Guide.
The content is directly extracted from the User Guide.
I'll extend it with aditional content more specific for
developers. The section "How application developers
can leverage AAA to provide servlet security" will be
removed from the AAA User Guide in a separate commit.

Change-Id: I1f60a641cc46944e56e96b090246588ff7b4f8b6
Signed-off-by: David <david.suarez.fuentes@ericsson.com>