Updated git submodules
Project: integration/test master c1556a332e93f18ebe81535013492bad73155688

distchanges: Add protections around patches under review

distchanges breaks for patches that are under review and not merged yet. This
is because the gerrit queries assume they are merged. When it breaks
the distchanges output stops abruptly and doesn't continue
processing all the projects. This patch prints the log and then keeps going.
There is another patch under review to get the gerrits for the unmerged patches,
so this patch here just keeps the logs going.

Change-Id: Ie451ccb083db3873f5680d78f1f8b0a36b0b9fc8
Signed-off-by: Sam Hague <shague@redhat.com>