Updated git submodules
Project: integration/test master 952b8b6943d6a3728e53599faf0e3497188e26c6

Check that DHCP server address is not None

in some cases it seems the nova console-log will already show the
obtained lease (the instance ip address) but not yet show the
nameserver string we are looking for (test moving too fast). So,
we need to make sure we keep checking until we have all of it.

Change-Id: If50742be982e54a8c9f0274de262ae24a2144faf
Signed-off-by: mor.zvi <mor.zvi@hpe.com>
Signed-off-by: Jamo Luhrsen <jluhrsen@redhat.com>